+++ Source/Lib/Fluxor.Blazor.Web.ReduxDevTools/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 ﻿using System.Runtime.CompilerServices;
 
+[assembly:InternalsVisibleTo("Fluxor.Blazor.Web.ReduxDevTools.UnitTests")]
 [assembly:InternalsVisibleTo("DynamicProxyGenAssembly2")]